extracts setup teardown and variables to BasePurchasesTest
